EZP-27015: Blocked deleting keyword if published version exists

+++ eZ/Publish/Core/FieldType/Keyword/KeywordStorage/Gateway/LegacyStorage.php
@@ -3,8 +3,9 @@
 namespace eZ\Publish\Core\FieldType\Keyword\KeywordStorage\Gateway;
 
 use eZ\Publish\Core\FieldType\Keyword\KeywordStorage\Gateway;
-use eZ\Publish\SPI\Persistence\Content\Field;
 use eZ\Publish\Core\Persistence\Database\DatabaseHandler;
+use eZ\Publish\SPI\Persistence\Content\Field;
+use eZ\Publish\SPI\Persistence\Content\VersionInfo;
 
 class LegacyStorage extends Gateway
 {
@@ -55,10 +56,11 @@ class LegacyStorage extends Gateway
     /**
      * Stores the keyword list from $field->value->externalData.
      *
-     * @param \eZ\Publish\SPI\Persistence\Content\Field
+     * @param \eZ\Publish\SPI\Persistence\Content\Field $field
      * @param mixed $contentTypeId
+     * @param \eZ\Publish\SPI\Persistence\Content\VersionInfo $versionInfo
      */
-    public function storeFieldData(Field $field, $contentTypeId)
+    public function storeFieldData(Field $field, $contentTypeId, VersionInfo $versionInfo)
     {
         if (empty($field->value->externalData) && !empty($field->id)) {
             $this->deleteFieldData($field->id);
@@ -66,7 +68,10 @@ class LegacyStorage extends Gateway
             return;
         }
 
-        $existingKeywordMap = $this->getExistingKeywords($field->value->externalData, $contentTypeId);
+        $existingKeywordMap = $this->getExistingKeywords(
+            $field->value->externalData,
+            $contentTypeId
+        );
 
         $this->deleteOldKeywordAssignments($field->id);
 
@@ -110,10 +115,11 @@ class LegacyStorage extends Gateway
      * Stores the keyword list from $field->value->externalData.
      *
      * @param mixed $fieldId
+     * @param \eZ\Publish\SPI\Persistence\Content\VersionInfo $versionInfo
      */
-    public function deleteFieldData($fieldId)
+    public function deleteFieldData($fieldId, VersionInfo $versionInfo = null)
     {
-        $this->deleteOldKeywordAssignments($fieldId);
+        $this->deleteOldKeywordAssignments($fieldId, $versionInfo);
         $this->deleteOrphanedKeywords();
     }
 
@@ -281,9 +287,17 @@ class LegacyStorage extends Gateway
         return $keywordIdMap;
     }
 
-    protected function deleteOldKeywordAssignments($fieldId)
+    protected function deleteOldKeywordAssignments($fieldId, VersionInfo $versionInfo = null)
     {
         $dbHandler = $this->getConnection();
+        // If current version being asked to be deleted is not published, then don't do that if
+        // there is some other version which is published (as keyword table is not versioned)
+        if ($versionInfo instanceof VersionInfo &&
+            $versionInfo->status !== VersionInfo::STATUS_PUBLISHED &&
+            $versionInfo->contentInfo->isPublished
+        ) {
+            return;
+        }
 
         $deleteQuery = $dbHandler->createDeleteQuery();
         $deleteQuery->deleteFrom(
